LSTM Notebook
This should run top to bottom correctly. I have it set to run the entire dataset but if your system is not robust it may have trouble, in which case you can experiment with 'reviews_by_state_trimmed' dataframe which contains about 344,000 reviews. If this is still too much, run cell 7 again to trim even further.